% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBRS_data.R
\docType{data}
\name{cBRS_data}
\alias{cBRS_data}
\title{cBRS dataset for use with Myths and Methodologies Reliability Papers}
\format{
An object of class \code{data.frame} with 168 rows and 36 columns.
}
\usage{
data(cBRS_data)
}
\description{
The time points  are NHS, LHS, MHS, and HHS.
Alpha coefficients are designated as alphaLF and alphaHF for the low and high frequency range respectively. Estimates are provided for the mean and median of the frequency ranges.
Transfer function estimates are designated by LF_TF and HF_TF for estimates in the low and high frequency ranges, respectively. Estimates are provided for the mean and median of the frequency ranges. Further, peak estimates in the frequency ranges are designated as PeakLF_TF and PeakHF_TF, respectively.
Gain is designated as the ‘\emph{TF}’ whereas Phase and Coherence are designated as ‘\emph{Phase}’ and ‘\emph{Coh}’, respectively.
Sequence estimates for combined up and down directional changes in SBP and RR is designated as ‘Combined_Slope’.
The baroreflex efficiency index is designated as ‘BEI’.
Time domain gain (i.e., GTIME) is designated as ‘TDG_mean’.
Descriptive statistics of mean and sd of the RR interval (‘RRint), DBP (‘DBP’), MAP (‘MAP’), and SBP (‘SBP’) time series are designated as ‘_mean’ and ‘_sd’ respectively.
}
\keyword{datasets}
