% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cBRS_squats_data.R
\docType{data}
\name{cBRS_squats_data}
\alias{cBRS_squats_data}
\title{cBRS squat dataset for use with Myths and Methodologies Reliability Papers}
\format{
An object of class \code{data.frame} with 84 rows and 25 columns.
}
\usage{
data(cBRS_squats_data)
}
\description{
The time points are PRE and POST.
Alpha coefficients are designated as alphaLF and alphaHF for the low and high frequency range respectively. Estimates are provided for the mean and median of the frequency ranges.
Transfer function estimates are designated by LF_TF and HF_TF for estimates in the low and high frequency ranges, respectively. Estimates are provided for the mean and median of the frequency ranges. Further, peak estimates in the frequency ranges are designated as PeakLF_TF and PeakHF_TF, respectively.
Gain is designated as the ‘\emph{TF}’ whereas Phase and Coherence are designated as ‘\emph{Phase}’ and ‘\emph{Coh}’, respectively.
}
\keyword{datasets}
