% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_power_plot_function.R
\name{coeff_power_plot_fun}
\alias{coeff_power_plot_fun}
\title{Title: Coefficient of variation and Power plots for different tools and effect sizes}
\usage{
coeff_power_plot_fun(
  MDC_input = NA,
  grand_mean,
  fig_type,
  SESOI = c(1, 3, 5, 10, 20, 25, 30),
  CoV = seq(4, 50, 2),
  CV1 = NA,
  CV2 = NA,
  all_percent = FALSE
)
}
\arguments{
\item{MDC_input}{vector of values referring to the minimum detectable change estimates. Only required for power curve}

\item{grand_mean}{numeric value for the grand mean. If coefficient figure then this is the value consistent for both measurement tools. If a power curve, it is the grand mean of the one tool.}

\item{fig_type}{string input taking form of "coefficient" or "power". Coefficient is to compare sample sizes across CV values. Power is to compare sample sizes across power for given effect sizes.}

\item{SESOI}{vector of values for the smallest effect size of interest (default: 1, 3, 5, 10, 20, 25, 30\%)}

\item{CoV}{vector of values for the coefficient of variation range important for the "coefficient" plot. Default is a sequence from 4 to 50 with increments of 2.}

\item{CV1}{numeric value for the coefficient of variation. If coefficient figure then the 1st CV would be the poorer tool, if power figure then CV for the only tool (and CV2 == NA)}

\item{CV2}{numeric value for the Coefficient of variation. If coefficient figure this is required as the better of the two tools.}

\item{all_percent}{boolean input. Default is FALSE. TRUE when the smallest effect size of interest and the MDC are the same unit (i.e., \%), so don't need to transfer to absolute values.}
}
\value{
figure either for coefficient of variation or power curve and estimating sample size.
}
\description{
NOTES: Function to generate plots for sample size estimations based on the coefficient of variation of two tools given a similar mean estimate
NOTES: Or alternatively a power plot for sample size estimations based on effect size and power.
}
\keyword{coefficient}
\keyword{coefficient,}
\keyword{correlation}
\keyword{intra-class}
\keyword{of}
\keyword{reliability,}
\keyword{variation}
