% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_extraction_function.R
\name{effectsize_table_function}
\alias{effectsize_table_function}
\title{Title: Extracting standardised effect sizes from data}
\usage{
effectsize_table_function(data_list, comparison, levels)
}
\arguments{
\item{data_list}{list of data tables}

\item{comparison}{string taking form of either "consecutive" or "incremental}

\item{levels}{numeric value for the number of comparisons (e.g., 2 levels is PRE and POST; 4 levels is NHS, LHS, MHS, HHS)}
}
\value{
output will be a table (dataframe) of the effect size calculations
}
\description{
Title: Extracting standardised effect sizes from data
}
\keyword{coefficient}
\keyword{coefficient,}
\keyword{correlation}
\keyword{intra-class}
\keyword{of}
\keyword{reliability,}
\keyword{variation}
