% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_req_sample_sizes_function.R
\name{formatting_req_samples_function}
\alias{formatting_req_samples_function}
\title{Title: Formatting sample size data and power plots}
\usage{
formatting_req_samples_function(df, NHS, LHS, MHS, HHS)
}
\arguments{
\item{df}{dataframe}

\item{NHS}{string input taking form of "Y" or "N"}

\item{LHS}{string input taking form of "Y" or "N"}

\item{MHS}{string input taking form of "Y" or "N"}

\item{HHS}{string input taking form of "Y" or "N"}
}
\value{
a dataframe with the number of participants required for each sample size and each effect size
}
\description{
NOTES:
In the case of sudomotor, the one level will be designated via NHS
In the case of cardiovascular it will be NHS and HHS
}
\keyword{coefficient}
\keyword{coefficient,}
\keyword{correlation}
\keyword{intra-class}
\keyword{of}
\keyword{reliability,}
\keyword{variation}
