% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_sample_size_function.R
\name{full_sample_size_function}
\alias{full_sample_size_function}
\title{Title: Convert MDC to standardised effects size and extract sample size data and power plot}
\usage{
full_sample_size_function(
  df,
  NHS,
  LHS,
  MHS,
  HHS,
  SESOI = c(1, 3, 5, 10, 20, 25, 30)
)
}
\arguments{
\item{df}{dataframe}

\item{NHS}{string input taking form of "Y" or "N"}

\item{LHS}{string input taking form of "Y" or "N"}

\item{MHS}{string input taking form of "Y" or "N"}

\item{HHS}{string input taking form of "Y" or "N"}

\item{SESOI}{vector of numeric values defining the smallest effect size of interest (as \%). Default is: 1, 3, 5, 10, 20, 25, 30\%}
}
\value{
list of figures (1st minor index) followed by sample size requirements (2nd and 3rd index)
}
\description{
NOTES: In the case of sudomotor, the one level will be designated via NHS
Use the plot_and_sizes function to generate the sample sizes required and the plots for those data (i.e., power curves)
}
\keyword{coefficient}
\keyword{coefficient,}
\keyword{correlation}
\keyword{intra-class}
\keyword{of}
\keyword{reliability,}
\keyword{variation}
