% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microvascular_data.R
\docType{data}
\name{microvascular_data}
\alias{microvascular_data}
\title{Microvascular function dataset for use with Myths and Methodologies Reliability Papers}
\format{
An object of class \code{data.frame} with 672 rows and 8 columns.
}
\usage{
data(microvascular_data)
}
\description{
An additional factor of “Laser” is presented to designate the laser placement on the dorsal aspect of the forearm.
Laser 1, 2, and 3 are individual laser units, whereas Laser 4 represents the average value of the three Lasers.
Skin blood flux, presented as an absolute value and as a value relative to the maximum achieved during combined
whole-body heating to 1.8°C and local heating to 44°C are designated as “SKBF” and “SKBF_REL”, respectively.
Cutaneous vascular conductance, presented as an absolute value and as a value relative to the maximum achieved during
combined whole-body heating to 1.8°C and local heating to 44°C are designated as “CVC” and “CVC_REL”, respectively.
}
\keyword{datasets}
