% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_sizes_function.R
\name{plot_and_sizes_function}
\alias{plot_and_sizes_function}
\title{Title: sample size calculation and power curve plot for Reliability Manuscripts}
\usage{
plot_and_sizes_function(input)
}
\arguments{
\item{input}{A vector of effect sizes is passed to the function (typically increments of 0.5)}
}
\value{
list of required sample sizes and figures
}
\description{
Title: sample size calculation and power curve plot for Reliability Manuscripts
}
\keyword{coefficient}
\keyword{coefficient,}
\keyword{correlation}
\keyword{intra-class}
\keyword{of}
\keyword{reliability,}
\keyword{variation}
