% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_extraction_function.R
\name{reliability_extraction_function}
\alias{reliability_extraction_function}
\title{Title: Extracting reliability Estimates from Raw Data}
\usage{
reliability_extraction_function(n, long, wide)
}
\arguments{
\item{n}{a numeric values indicating the number of variables in the dataframe}

\item{long}{a dataframe in long-form}

\item{wide}{a dataframe in wide-form}
}
\value{
list of calculated reliability statistics (hopkins) and icc (psych) and raw data for effect size calculations
}
\description{
This function will take a value of the number of variable required, and a long and wide format of the data frame
Requires the reliability_function from the Hopkins spreadsheet and also requires the psych package for ICC
}
\keyword{coefficient}
\keyword{coefficient,}
\keyword{correlation}
\keyword{intra-class}
\keyword{of}
\keyword{reliability,}
\keyword{variation}
