% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudomotor_data.R
\docType{data}
\name{sudomotor_data}
\alias{sudomotor_data}
\title{Sudomotor function dataset for use with Myths and Methodologies Reliability Papers}
\format{
An object of class \code{data.frame} with 168 rows and 63 columns.
}
\usage{
data(sudomotor_data)
}
\description{
Nine body sites are designated specifically as:
Abdomen, Bicep, Calf, Chest, Foot, Forearm, Forehead, Hand, and Quad.
Three regional groupings are designated specifically as:
Trunk, Arm, and Leg
Sweat rate is designated as “SR” at the end of the site or region name.
Sweat rate normalised to NHS is designated as “SR_norm’ at the end of the site or region name.
Temperature onset of sweating is designated as “ONSET” at the end of the site or region name.
Onset of sweating related to relative change in mean body temperature is designated as “ONSET_rel” at the end of the site or region name.
Thermosensitivity of sweating is designated as “SENS” at the end of the site or region name.
}
\keyword{datasets}
