% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_table_function.R
\name{systematic_table_function}
\alias{systematic_table_function}
\title{Title: extract systematic differences for reliability data}
\usage{
systematic_table_function(data_list, NHS, LHS, MHS, HHS, levels)
}
\arguments{
\item{data_list}{list of data tables}

\item{NHS}{string input taking form of "Y" or "N"}

\item{LHS}{string input taking form of "Y" or "N"}

\item{MHS}{string input taking form of "Y" or "N"}

\item{HHS}{string input taking form of "Y" or "N"}

\item{levels}{numeric value indicating number of heating levels (2: NHS and HHS; 4: NHS, LHS, MHS, and HHS)}
}
\value{
list of figures (1st minor index) followed by sample size requirements (2nd and 3rd index)
}
\description{
Title: extract systematic differences for reliability data
}
\keyword{coefficient}
\keyword{coefficient,}
\keyword{correlation}
\keyword{intra-class}
\keyword{of}
\keyword{reliability,}
\keyword{variation}
